% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dmvnrm_arma}
\alias{dmvnrm_arma}
\title{Internal function to calculate the density of multivariate normal distribution.}
\usage{
dmvnrm_arma(x, mean, sigma, logd = FALSE)
}
\description{
Internal function to calculate the density of multivariate normal distribution.
}
\keyword{internal}

