% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mvrnormArma}
\alias{mvrnormArma}
\title{Internal function to sample from multivariate normal distribution.}
\usage{
mvrnormArma(n, mu, sigma)
}
\description{
Internal function to sample from multivariate normal distribution.
}
\keyword{internal}

