% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{polyurncpp}
\alias{polyurncpp}
\title{Internal function to sample cluster membership indicator}
\usage{
polyurncpp(e, muA0, sigma0, A, sigma, sigmaInv, sigma0Inv, c)
}
\description{
Internal function to sample cluster membership indicator
}
\keyword{internal}

