% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_DP_fun.R
\name{samMu}
\alias{samMu}
\title{Internal function to sample cluster-specific means of eta}
\usage{
samMu(e, eta, mu0, sigma0Inv, sigmaInv)
}
\description{
Internal function to sample cluster-specific means of eta
}
\keyword{internal}

