% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_DP_fun.R
\name{samSig}
\alias{samSig}
\title{Internal function to sample gene-specific variances}
\usage{
samSig(v1, v2, A, lam, eta)
}
\description{
Internal function to sample gene-specific variances
}
\keyword{internal}

